Rename model_cache_default.py -> model_cache.py.
